requirements.txt
pandas==2.3.1
matplotlib==3.9.4
seaborn==0.13.2
numpy==2.0.2
sqlalchemy==2.0.41
pymysql==1.1.1
mysql-connector-python==9.3.0
openpyxl==3.1.5
jupyter==1.1.1
ipykernel==6.29.5
nbconvert==7.16.4
